move state_of_mark
