+++ autogen/screen_parsing/browser_utils/__init__.py
@@ -4,7 +4,7 @@ from .selenium_markdown_browser import SeleniumMarkdownBrowser
 from .playwright_markdown_browser import PlaywrightMarkdownBrowser
 from .markdown_search import AbstractMarkdownSearch, BingMarkdownSearch
 from .mdconvert import MarkdownConverter, UnsupportedFormatException, FileConversionException, DocumentConverterResult
-
+from .state_of_mark import add_state_of_mark
 
 __all__ = (
     "AbstractMarkdownBrowser",
@@ -17,4 +17,5 @@ __all__ = (
     "UnsupportedFormatException",
     "FileConversionException",
     "DocumentConverterResult",
+    "add_state_of_mark",
 )
